Merged PR 4: Merge russellseymoour/build-manifest to master
Created new build process so that it is configurable.

A new script `build.ts` (`build.js` when compiled) is used to drive all aspects of the build for the extension. The accompanying `build.json` file in the root directory controls where the extensions is built. It also contains the files that need to be packaged into each task.

There are a number of options that can be set on the build process. All of them have valid defaults that should mean the build can be run without any options.

Running the command `node scripts/build.js --help` will show the defaults for each option.