Merged PR 5: TypeScript Linting
TypeScript linting has been added to the build process.

It will check all TypeScript files in the `scripts` directory and the TypeScript files for a task. It will _not_ descend into the node_modules directory of a task if it is exists.

The scripts are now compiled into the `bin` dir so that they do not pollute the source when running locally.