Merged PR 6: Extension should run on Windows
All tasks will now run on Windows based agents. I have not yet tested if it works on Hosted agents.

As much as possible the detection logic for the platforms has been put into one class, `builtin`. This responsible for getting the paths fpor the various commands and normalising them. In some tasks the values _maybe_ overwritten by what has been specified in the Advanced options of the task.

Two new tasks have been added
 - Install ChefDK
 - Install InSpec

These tasks check the environment that the agent is running in and will determine if it has privileges to perform the installation. So on Linux this is either running as root or has the ability to run `sudo`. On Windows this is checking that it is running with elevated privileges.

The tasks that uses the applications no longer attempt to install the software if it does not exist. This is because I was having issues with Promises in `nodejs`. I will revisit this at a later date.