Merged PR 7: Tidied up internal libraries and added Knife tasks
The two internal libraries `builtin` and `inputs` have bene replaced by `settings`. This new library handles the task inputs and normalization of paths on different platforms. The output is more or less the same, just that there is now fewer libraries to call.

Two new tasks have been added:

**Install Gem**

Installs any gem within the ChefDK context, e.g. `chef gem install xxxx`. The primary use case for this is to install knife plugiuns.
The task will check to see if the gem is installed, or has been force installed by the options, before installation is attempted.

Once ChefDK is installed this task does not require any addiaionl privileges.

**Execute Knife**

This task allows `knife` to be run. It is designed to be used in conjunction with the prvious task. For example a request has been made to allow cookbooks to be uploaded to Artifactory. This is a knife plugin called `knife-art`. However we do not want to bundle this into the extension. So the previous task allows for this to be installed and this task allows it to be executed.

So that endpoint credentials can be reused there is the option of selecting an endpoint, but it is not required.

The task will always run `knife`. The options state what command is to be run and any options that are to be specified.

There are 3 tokens which can be added to the arguments that will be replaced by the values in the endpoint (if selected):

| Token        | Description                                                      |
|--------------|------------------------------------------------------------------|
| `{URL}`      | URL of the specified endpoint. Does not have to be a Chef server |
| `{USERNAME}` | Username as specified in the endpoint                            |
| `{PASSWORD}` | The password (or client key) as specified in the endpoint.       |

Additionally the 'Private Key' task option tells the task that the private key should be written out to a file on disk and then the `--key` option appended to the knife arguments.

Finally if the endpoint has the SSLVerify option set to false, then a temporary `knife.rb` file is written out with the contents of `ssl_verify_mode :verify_none`. This is then appended to the knife arguments using `--config`.

By providing this task it is allows people to perform more tasks without the need for us to explicitly support them.