Merged PR 8: Added TK and Linting tasks to the extension
3 new tasks have been adde to the extesion

**Install Cookbook Gems**
This is a handy wrapper task for `chef exec bundle` that installs the Gems required for testing for the cookbook.
The task allows the path to be set for the `gemfile` allowing cookbooks to be stored in a different folder within the source. (Default: `$(Build.SourcesDirectory)`)

**Linting**
This task allows 3 types of linting. It assumes that the tasks are setup in a `Rakefile`. As with the previous task this is configurable on the task. (Default: `$(Build.SourcesDirectory)`)
The three designated actions are:
 - Chefstyle
 - Foodcritic
 - Chefspec
Each of the actions can be overriden by specifying the `Rakefile` task to call in the advanced parameters for the task.

**Test Kitchen**
This task allows test kitchen tests to be performed. It is assumed that these tests will be based on the Kitchen Azurerm driver (which can be installed using the Install Gem task). The credentials for this are set using the Azure Resource Manager endpoint.
The default file to look for is `.kitchen.yml` but this can be overridden on the task parameters.

The tasks have been cleaned up so there are no unnecessary tasks being packaged. This has reduced the overall size of the extension so it can be deployed.
Have used the built in methods for filesystem tasks within the `tl` libray and removed the depdency on `fs-extra`.