Merged PR 9: Knife Exec Task Issues
Corrected the issues that were causing the Knife Exec task to fail.

This was down to the refactoring from the last change and some of the variable names were not changed.
Additionally the arguments for knife were not being appended correctly so the whole argument list was being globbered in certain situations.