Use knife supermarket instead of knife cookbook site
Several years ago we started the process of replacing knife cookbook
site with knife supermarket. In Chef 15 we introduce a deprecation
warning when knife cookbook site is used and in Chef 16 we will remove
the knife cookbook site command entirely. This avoids the deprecation
warning by using the correct command.

Signed-off-by: Tim Smith <tsmith@chef.io>